fix: add client-side logic to server action
